Updating creating method
